Isolate Period scalar tests
